Support mixWithOthers on iOS
The mixWithOthers option on iOS allows apps to specify that audio from
other sessions should be mixed in with the current app's session, even
in categories such as Playback.  We should support this option in the
library.

* Update the setCategory function to support an additional parameter
which sets the requisite withOptions parameter on setCategory.
* Remove the key parameter from setCategory since it wasn't being used.
* Add a static setCategory function in the JS.  For backwards
compatibility, the old version just calls this new function.  Also add
a guard so that android doesn't get borked.
* Remove the automatic call to enable since it gets called before the
category has been set up (which may lead to side effects), and is not
needed.
* Update docs.